fix: cast columnar weighted output to result dtype before index_add_
src_weights * col_out produces float32 when routing weights are
computed in float32 but result is bfloat16, causing index_add_ to
fail with a dtype mismatch. Cast to result.dtype before accumulation.

+++ coral/models/columnar.py
@@ -118,7 +118,7 @@ class ColumnarTransformerBlock(nn.Module):
             result.index_add_(
                 0,
                 src_samples,
-                src_weights.unsqueeze(-1).unsqueeze(-1) * col_out,
+                (src_weights.unsqueeze(-1).unsqueeze(-1) * col_out).to(result.dtype),
             )
 
         return result, routing_logits
